Updated tests; new screenshots; formatting
